separate module for old token nonces
