+++ locked-asset/simple-lock-energy/src/unlock_with_penalty.rs
@@ -33,10 +33,10 @@ pub trait UnlockWithPenaltyModule:
     + simple_lock::locked_token::LockedTokenModule
     + simple_lock::token_attributes::TokenAttributesModule
     + elrond_wasm_modules::default_issue_callbacks::DefaultIssueCallbacksModule
-    + crate::token_whitelist::TokenWhitelistModule
     + crate::util::UtilModule
     + crate::energy::EnergyModule
     + crate::lock_options::LockOptionsModule
+    + crate::old_token_nonces::OldTokenNonces
     + crate::events::EventsModule
     + elrond_wasm_modules::pause::PauseModule
 {
